Refactored Home and Applications into Portfolio. Navigation functional, new icons chosen.
